nrf_desktop: replace use of removed aliases
The pwm-0 and pwm-1 aliases are gone now that we can
refer to nodes by node label. That is, we need to use
DT_NODELABEL(pwmX) instead of DT_ALIAS(pwm_X), since the pwm_X alias
is gone. It only ever existed because the pwmX form could not be
referred to in the legacy devicetree macros.

The same is true of the adc-0 alias.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>