nrf_cloud: handle constification in struct mqtt_utf8
In Zephyr commit 8c29aa9881dc459e52288817158fb2f4e6ee62b5, struct
mqtt_utf8's "utf8" member was made const. This requires us to cast
away const in the free routine, which is a macro wrapping k_free(),
whose argument is not const, to avoid a discarded-qualifiers warning.
From a cursory analysis of the call graph, this appears to be a safe
thing to do.

In a similar vein, introduce a non-const pointer to a value which is
both assigned to a utf8 member and passed as the snprintf() buffer.
The existing error handling in this case is a little suspect: it
leaves size at a nonzero value after setting the pointer to NULL, and
logs but does not return an error.

It would seem to be better to set size to 0 here before returning;
I'll do that in a separate patch.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>