nrf_desktop: adjust seemingly unsafe macro use
While this seems to work, using #if within the arguments to a variadic
macro as is done in hids.c seems a bit sketchy. Switch this up to use
IF_ENABLED, which is semantically equivalent but should nest properly
within a macro expansion regardless of how BT_GATT_HIDS_DEF is defined.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>