bluetooth: mesh: settings_set callback handling hack
This is just a temporary measure to handle the parameter list change
to the corresponding callback. I don't know how to support this
feature when name is not NULL, so just return an error in that case.
I'll leave it to the mesh maintainers to take it from here or provide
some other advice.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>