lwm2m_carrier: handle change to logging internals
Upstream Zephyr commit c73578f37ac7174f362be6865dc8307e4b324932 added
a parameter to the internal log_generic() routine which specifies the
behavior when it comes to duplicating strings pointed to by varargs
corresponding to %s in the format.

Since there is a lwm2m_os_log_strdup(const char *str) function and
this appears to have been added for OpenThread, I'm guessing that the
intent for lwm2m_carrier is that callers are expected to strdup their
strings as needed, so passing LOG_STRDUP_SKIP ought to be the right
thing to do.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>